requirements.txt
# general  modules
django==4.2
Pillow
djangorestframework

# env controlls
python-decouple

# third party modules
markdown       
django-filter  
drf-spectacular
djangorestframework-simplejwt
django-cors-headers
django-robots
django-debug-toolbar

# deployment module
gunicorn

# refactoring and tests
black
flake8
pytest
pytest-django
Faker

# background proccess and cache
celery
django-redis

# database client
psycopg[binary,pool]